WIP checkpoint: promote the durable spine fake behind test-support
Move FakeSpineStore, its state and marker types, the bounded rendezvous and
the rollout fence fixture builders out of kin-spine's private test module
into a test_support module gated on test-or-test-support, so a consumer can
drive the real FirestoreSpineBackend over a fake transport instead of
writing a second implementation of the CAS semantics.

Not yet compiled; import cleanup follows.
